Journal Crash Fixed API Key Changes for added in UserResponse also Key Changes added for Past Report sleep value in UpdatedModule RESPONSE sleepDuration was int now  changed to double face scan click check according new key from backend Ai report click using new key from backend

+++ app/src/main/java/com/jetsynthesys/rightlife/newdashboard/model/UpdatedModule.kt
@@ -13,7 +13,7 @@ data class UpdatedModule(
     @SerializedName("burned") val burned: String?,
     @SerializedName("mindfulTime") val mindfulTime: String?,
     @SerializedName("calories") val calories: String?,
-    @SerializedName("sleepDuration") val sleepDuration: Int?,
+    @SerializedName("sleepDuration") val sleepDuration: Double?,
     @SerializedName("isSelectedModule") val isSelectedModule: Boolean?,
     @SerializedName("mindfulnessMinutes") val mindfulnessMinutes: String?,
     @SerializedName("wellnessDays") val wellnessDays: String?,
